refactor!: rename project to blink-ripgrep.nvim
Issue
=====

It's confusing that this project has the same name as the original
project. This causes confusion and is not in anyone's best interest.

Solution
========

Rename the project to `blink-ripgrep.nvim`.
