+++ integration-tests/cypress/e2e/cmp-rg/health_spec.cy.ts
@@ -5,8 +5,8 @@ describe("the healthcheck", () => {
       // wait until text on the start screen is visible
       cy.contains("If you see this text, Neovim is ready!")
 
-      cy.typeIntoTerminal(":checkhealth blink-cmp-rg{enter}")
-      cy.contains("OK blink-cmp-rg")
+      cy.typeIntoTerminal(":checkhealth blink-ripgrep{enter}")
+      cy.contains("OK blink-ripgrep")
       cy.contains("WARN").should("not.exist")
     })
   })
