+++ package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@blink-cmp-rg/monorepo",
+  "name": "@blink-ripgrep/monorepo",
   "version": "1.0.0",
   "private": true,
   "license": "MIT",
