+++ spec/blink-ripgrep/rg-output.jsonl
